Refactor address counter functions
