+++ apps/block_scout_web/lib/block_scout_web/controllers/api/v2/stats_controller.ex
@@ -4,6 +4,7 @@ defmodule BlockScoutWeb.API.V2.StatsController do
   alias BlockScoutWeb.API.V2.Helper
   alias BlockScoutWeb.Chain.MarketHistoryChartController
   alias Explorer.{Chain, Market}
+  alias Explorer.Chain.Address.Counters
   alias Explorer.Chain.Cache.Block, as: BlockCache
   alias Explorer.Chain.Cache.{GasPriceOracle, GasUsage}
   alias Explorer.Chain.Cache.Transaction, as: TransactionCache
@@ -43,7 +44,7 @@ defmodule BlockScoutWeb.API.V2.StatsController do
       conn,
       %{
         "total_blocks" => BlockCache.estimated_count() |> to_string(),
-        "total_addresses" => @api_true |> Chain.address_estimated_count() |> to_string(),
+        "total_addresses" => @api_true |> Counters.address_estimated_count() |> to_string(),
         "total_transactions" => TransactionCache.estimated_count() |> to_string(),
         "average_block_time" => AverageBlockTime.average_block_time() |> Duration.to_milliseconds(),
         "coin_price" => exchange_rate.usd_value,
