+++ apps/block_scout_web/lib/block_scout_web/templates/address/_tabs.html.eex
@@ -8,7 +8,7 @@
         class: "card-tab #{tab_status("transactions", @conn.request_path)}",
         to: AccessHelper.get_path(@conn, :address_transaction_path, :index, @address.hash)
       ) %>
-  <%= if Chain.check_if_token_transfers_at_address(@address.hash) do %>
+  <%= if Counters.check_if_token_transfers_at_address(@address.hash) do %>
     <%= link(
         gettext("Token Transfers"),
         class: "card-tab #{tab_status("token-transfers", @conn.request_path)}",
@@ -16,7 +16,7 @@
         to: AccessHelper.get_path(@conn, :address_token_transfers_path, :index, @address.hash)
         ) %>
   <% end %>
-  <%= if Chain.check_if_tokens_at_address(@address.hash) do %>
+  <%= if Counters.check_if_tokens_at_address(@address.hash) do %>
   <%= link(
         gettext("Tokens"),
         class: "card-tab #{tab_status("tokens", @conn.request_path)}",
@@ -24,7 +24,7 @@
         "data-test": "tokens_tab_link"
       ) %>
   <% end %>
-  <%= if Chain.check_if_withdrawals_at_address(@address.hash) do %>
+  <%= if Counters.check_if_withdrawals_at_address(@address.hash) do %>
   <%= link(
         gettext("Withdrawals"),
         class: "card-tab #{tab_status("withdrawals", @conn.request_path)}",
@@ -44,14 +44,14 @@
         "data-test": "coin_balance_tab_link",
         to: AccessHelper.get_path(@conn, :address_coin_balance_path, :index, @address.hash)
       ) %>
-  <%= if Chain.check_if_logs_at_address(@address.hash) do %>
+  <%= if Counters.check_if_logs_at_address(@address.hash) do %>
     <%= link(
         gettext("Logs"),
         class: "card-tab #{tab_status("logs", @conn.request_path)}",
         to: AccessHelper.get_path(@conn, :address_logs_path, :index, @address.hash)
         ) %>
   <% end %>
-  <%= if Chain.check_if_validated_blocks_at_address(@address.hash) do %>
+  <%= if Counters.check_if_validated_blocks_at_address(@address.hash) do %>
     <%= link(
       gettext("Blocks Validated"),
       class: "card-tab #{tab_status("validations", @conn.request_path)}",
