+++ apps/explorer/lib/explorer/counters/addresses_with_balance_counter.ex
@@ -7,7 +7,7 @@ defmodule Explorer.Counters.AddressesWithBalanceCounter do
 
   use GenServer
 
-  alias Explorer.Chain
+  alias Explorer.Chain.Address.Counters
 
   @table :addresses_with_balance_counter
 
@@ -104,7 +104,7 @@ defmodule Explorer.Counters.AddressesWithBalanceCounter do
   Consolidates the info by populating the `:ets` table with the current database information.
   """
   def consolidate do
-    counter = Chain.count_addresses_with_balance()
+    counter = Counters.count_addresses_with_balance()
 
     insert_counter({cache_key(), counter})
   end
